requirements.txt
google-cloud-pubsub
google-cloud-asset